+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes
 * Remove [Rome](https://github.com/tmspzz/Rome) support
 * Change that the altool use the apiKey and apiIssue. (username and password should still work)
 * Change that app extensions gets only the keychain-access-group of the main bundle set.
+* Updated appstore upload task so that it works with Xcode13
 
 # 0.20.1 (2. Dec, 2020)